Refactor map animation code out and drastically optimize
